Add custom volumes to the replica initContainers.
In the EnforceSpec method, we now append the custom volumes to the
first initContainer, if there is at least one initContainer.

In the removeCustomVolumeMounts method, we now remove custom volumes
from the first initContainer, if there is at least one initContainer.

It is assumed in most cases that the initContainer and the runtime
container will have the same custom volumes for the sake of checking
the spec, but in the places where spec is enforced they are checked
individually.

This would resolve reactive-tech/kubegres#122.